requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0
psycopg2-binary>=2.9
gunicorn>=21.2
python-barcode>=0.15
Pillow>=10.0
